Reset all migrations in core and credential app.
