+++ credentials/apps/core/models.py
@@ -1,12 +1,36 @@
 """ Core models. """
 
 from django.contrib.auth.models import AbstractUser
+from django.contrib.sites.models import Site
 from django.db import models
 from django.utils.translation import ugettext_lazy as _
 
 
+class SiteConfiguration(models.Model):
+    """
+    Site configuration model for custom credentials sites.
+    """
+    site = models.OneToOneField(Site, null=False, blank=False)
+    lms_url_root = models.URLField(
+        verbose_name=_('LMS base url for custom site'),
+        help_text=_("Root URL of this site's LMS (e.g. https://courses.stage.edx.org)"),
+        null=False,
+        blank=False
+    )
+    theme_scss_path = models.CharField(
+        verbose_name=_('Path to custom site theme'),
+        help_text=_("Path to site theme's main SCSS file"),
+        max_length=255,
+        null=False,
+        blank=False
+    )
+
+    def __unicode__(self):
+        return unicode(self.site.name)
+
+
 class User(AbstractUser):
-    """Custom user model for use with OpenID Connect."""
+    """ Custom user model for use with OpenID Connect. """
     full_name = models.CharField(_('Full Name'), max_length=255, blank=True, null=True)
 
     @property
